requirements.txt
# Web & API
fastapi
uvicorn[standard]
python-multipart

# Data Mining & Processing
pandas
numpy
beautifulsoup4
requests
sqlalchemy

# AI & NLP
scikit-learn
openai
spacy
python-dotenv
